Desktop/ai-knowledge-hub/backend/requirements.txt
# Django + ASGI + REST
Django==4.2.14
djangorestframework==3.15.2
uvicorn[standard]==0.30.1
django-cors-headers==4.4.0
whitenoise==6.7.0

# Auth (JWT)
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0

# MongoDB
pymongo[srv]==4.8.0

# Realtime / sockets
python-socketio==5.11.4
python-engineio==4.4.0

# Optional AI / embeddings (install if using provider/local models)
openai==1.51.0
sentence-transformers==3.0.1
numpy==1.26.4

# Document parsing (optional)
pdfminer.six==20231228
python-docx==1.1.2

# Env
python-dotenv==1.0.1
